+++ src/hooks/useCartItem.ts
@@ -13,6 +13,7 @@ const useCartItems = () => {
       setLoading(true);
       try {
         const cartItems = await fetchItems();
+        console.log(cartItemIds);
 
         const newCartItemIds = cartItems.map((cartItem) => cartItem.product.id);
         setCartItem(newCartItemIds);
